requirements.txt
asgiref
Django==4.2
sqlparse
psycopg2-binary
pillow
ipython
pytils
django-crispy-forms==1.14.0
redis
python-dotenv~=1.0.1
APScheduler~=3.10.4
django_apscheduler
django-bootstrap-datepicker==1.5.0
django-bootstrap-datepicker-plus==5.0.5